corrected ambiguous overload of std::sqrt() in CImg.h

+++ src/CImg.h
@@ -14207,7 +14207,7 @@ namespace cimg_library {
             for (int k = i; k<height(); ++k) scale+= cimg::abs(U(i,k));
             if (scale) {
               for (int k = i; k<height(); ++k) { U(i,k)/=scale; s+= U(i,k)*U(i,k); }
-              f = U(i,i); g = (t)((f>=0?-1:1)*std::sqrt(s)); h=f*g-s; U(i,i) = f-g;
+              f = U(i,i); g = (t)((f>=0?-1:1)*(double)std::sqrt((double)s)); h=f*g-s; U(i,i) = f-g;
               for (int j = l; j<width(); ++j) {
                 s = 0;
                 for (int k=i; k<height(); ++k) s+= U(i,k)*U(j,k);
@@ -14224,7 +14224,7 @@ namespace cimg_library {
             for (int k = l; k<width(); ++k) scale+=cimg::abs(U(k,i));
             if (scale) {
               for (int k = l; k<width(); ++k) { U(k,i)/= scale; s+= U(k,i)*U(k,i); }
-              f = U(l,i); g = (t)((f>=0?-1:1)*std::sqrt(s)); h = f*g-s; U(l,i) = f-g;
+              f = U(l,i); g = (t)((f>=0?-1:1)*(double)std::sqrt((double)s)); h = f*g-s; U(l,i) = f-g;
               for (int k = l; k<width(); ++k) rv1[k]=U(k,i)/h;
               for (int j = l; j<height(); ++j) {
                 s = 0;
